Refactor: Moved Sponsors components to components [ _componet/sponsors was not deleted ]
