+++ app/page.jsx
@@ -3,9 +3,10 @@ import Eligibility from "./_components/Home/Eligibility";
 import Gallery from "./_components/Home/Gallery";
 import Hero from "./_components/Home/Hero";
 import PastWinners from "./_components/Home/PastWinners";
-import Sponsor from "./_components/Home/Sponsor";
+//import Sponsor from "./_components/Home/Sponsor";
+import Sponsor from "@/components/Home/Sponsors/sponsors";
 import Timeline from "./_components/Home/Timeline";
-import SponsorDev from "./_components/Home/SponsorDv";
+
 //
 export default function Home() {
   return (
@@ -15,9 +16,6 @@ export default function Home() {
       <Eligibility />
       <Timeline />
       <Gallery />
-      <SponsorDev />
-      
-      <Sponsor />
       <Sponsor />
       <PastWinners />
     </main>
